# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounts": {
      "post": {
        "description": "Creates an NFC-backed account for a user. Requires an admin API token.",
        "operationId": "AccountsController_createAccount_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create account",
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "description": "Returns a paginated list of accounts with each account owner (user) summary. Admins may filter by owner user id, owner user type, lock state, and a case-insensitive search on NFC UID or owner name fields.",
        "operationId": "AccountsController_listAccounts_v1",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Return only accounts owned by this user.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "ownerType",
            "required": false,
            "in": "query",
            "description": "Return only accounts whose owner has this user type.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "provider",
                "admin"
              ]
            }
          },
          {
            "name": "isLocked",
            "required": false,
            "in": "query",
            "description": "Return only locked (true) or unlocked (false) accounts.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on NFC card UID, owner display name, or owner login name (userName).",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 50, max 100).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip (pagination).",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAccountsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List accounts",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{accountId}/transactions": {
      "get": {
        "description": "Returns transactions where the account is the payer, payee, or mint recipient. The account owner or an admin may call this.",
        "operationId": "AccountsController_getAccountTransactions_v1",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "How many transactions to return (default 10, max 100).",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller is not the account owner or an admin"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List account transactions",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classurency API",
    "description": "API documentation for the Classurency backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "token",
        "type": "http"
      }
    },
    "schemas": {
...
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Transaction ID"
          },
          "type": {
            "type": "string",
            "enum": [
              "MINT",
              "PURCHASE"
            ],
            "description": "Transaction kind"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount"
          },
          "description": {
            "type": "object",
            "description": "Description",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "fromAccountId": {
            "type": "object",
            "description": "Source account ID (payer); null when not applicable (e.g. mint)",
            "nullable": true
          },
          "toAccountId": {
            "type": "object",
            "description": "Destination account ID",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "amount",
          "createdAt"
        ]
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "format": "uuid",
            "description": "User id."
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateAccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User display name"
          },
          "nfcCardUid": {
            "type": "string",
            "description": "NFC card UID assigned to the account"
          },
          "publicKeyHex": {
            "type": "string",
            "description": "Account public key (hex)"
          },
          "pin": {
            "type": "string",
            "description": "4-digit PIN for the NFC card"
          },
          "encryptedPrivateKeyForCard": {
            "type": "string",
            "description": "Encrypted private key for programming onto the card (one-time delivery)."
          }
        },
        "required": [
          "id",
          "name",
          "nfcCardUid",
          "publicKeyHex",
          "pin",
          "encryptedPrivateKeyForCard"
        ]
      },
      "AccountListOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "user",
              "provider",
              "admin"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "userName",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListAccountItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "nfcCardUid": {
            "type": "object",
            "nullable": true,
            "description": "NFC card UID when assigned."
          },
          "publicKeyHex": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "description": "Account balance."
          },
          "isLocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "description": "The user that owns this account.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountListOwnerDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "publicKeyHex",
          "balance",
          "isLocked",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
      "ListAccountsResponseDto": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListAccountItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters (ignoring limit/offset)."
          }
        },
        "required": [
          "accounts",
          "total"
        ]
      },
...
    }
  }
}